Fix DbChangeNotificationInternalTestModel setup failure caused by the patch for bug 237545

+++ foundation/eclipselink.core.test/src/org/eclipse/persistence/testing/models/employee/domain/Child.java
@@ -15,9 +15,10 @@
 package org.eclipse.persistence.testing.models.employee.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class Child implements java.io.Serializable{
-    public int id;
+    public BigDecimal id;
     /** Direct-to-field mapping, String -> VARCHAR. */
     public String firstName;
